requirements.txt
aiogram==3.0.0b7
pydantic>=2.0.0
python-dotenv>=1.0.0